Convert to event-based database and asn scheamas  - Update database schema to record events for datapoints, and  new/lost habs and node  - Rewrite database accessors for new schema  - Metadata inserts return the rowid, so cacheing can occur  - Add db check test suite  - Convert bdm db-cache to be event-wise
 - Update ASN protocol, and associated message generators and parsers
 - Add bionet_event_t datastructure used in libbionet and libbdm, and add to hab_t, no
 - Add uid to node, and use to identifiy nodes in bdm
 - Rename bdm_db_batch to bdm_db_batch_t
 - Add maximum commit attempt count

+++ data-manager/server/subscription.h
@@ -11,7 +11,7 @@
 typedef GHashTable bdm_peer_states_t;
 
 typedef struct {
-    int curr_seq; // -1 for unset. This means the subscription is in the pending state..
+    long curr_seq; // -1 for unset. This means the subscription is in the pending state..
 } bdm_subscription_t;
 
 bdm_peer_states_t * bdm_subscriptions_new(void);
